Update README.md
It should be "air gap" rather than "Air Gap" because it is not a proper noun (e.g. branded name).

Also, it's best to be consistent and use "you" instead of "we" throughout. Also, technical documentation typically avoids "we", so I suggest using "you" instead. But either way, consistency is ideal.

Looks good overall! It was a fun assignment to install this, and I was successful!