Adds a template for a shell + cluster lab
TL;DR
-----

Providse a templates track for a shell running independently of the K8s cluster

Details
-------

My first attempts at setting up a lab where the shell for running CLI commands is
independent of the Kubernetes cluster they are run against ran into issues with
the way that the Instruqt bootstrap injects SSH keys and the Dropbear server. This
change captures the pattern I used to workaround the limitations that caused my
earlier lifecycle scripts to fail.

When configurating a track with this setup, the best way to assure access to the
K8s cluster is to take the default config from `/etc/rancher/k3s/k3s.yaml` and
use it from shell host in the sandbox. My first attempts at this ran into two
issues:

1. The timing of `/etc/rancher/k3s/k3s.yaml` being available.  The best way
   around this was to push the file from the cluster host onto the shell host
   so that I could assure the file was in place.
2. A misalignment of SSH key types between the Dropbear server (on the shell
   host) and the SSH client on the cluster host. These are both injected by
   the Instruqt bootstrap. I consider the behavior a bug, but I was able to
   find a workaround.

The workaround gets around the bootstrap issue by assuring that the SSH server
on the shell host has an RSA key to present, then making sure that the SSH client
on the cluster accepts it. Both changes are necessary becuause the bootstrap
was leaving the two in a state where they cluster could not connect to the shell
with it's RSA id and the server only having an ECDSA key.

With this is place, I push the file from the cluster host to the shell host
during track setup, then update it with the correct address when the shell host
sets up. The shell setup also makes sure that the user `replicant` can access
its own Kubernetes config. This step is needed since the edits in the setup
script are run as root.