Implements initial airgap lab for an embedded cluster (#191)
TL;DR
-----

Moves the full airgap lab to Instruqt

Details
-------

Implements a lab showing key features of the airgap workflow using
the Instuqt platform. This lab migrates "Lab 05" from the earlier
implementation so we can show a completed airgap scenairo using
Instruqt.

The Lab is implemented with two sandbox instances and a gcloud
project. One sandbox instance is a worker for setting up the airgap
and is not used directly in any challenge. This shouldn't be needed,
but there were some timing issues in the Instruqt bootstrap that
made it easier to bootstrap from a container sandbox than VM 
sandbox.

The second sandbox instance is the Jumpbox used by the student to 
complete the Lab. This has to be a VM instead of a conatiner so we 
can direct traffic to it from teh leaner's browser after setting up
port forawrding for the Admin Console. We may not need to do this 
long term, since we asked the Admin Console team to make some 
changes that would allow using an Instrqut service instead.

### Using `tmux`

This is the first lab I've built using `tmux` to preserve the shell
environment between challenges. The shell for the `jumpbox`
image starts or joins a `tmux` session which uses `su` so the
learner is not root. Some challenge cleanup scripts clear the 
scrollback buffer to give a clean look like challenges in tracks
that don't share a shell. Other don't blank it, like when the 
KOTS bundle is downloading, so that the learner can keep 
important context between challenges.

### Jumpbox Image

The Jumpbox image is created with Packer using the template in
`packer/jumpbox`, while the lab is implemented in `instruqt/airgap-embedded`.
Like the `shell` docker image, I took on a small bit of technical
debt and haven't implemented CI/CD for the jumpbox machine image
as part of getting this lab built. I've also updated the airgap
template to use the Jumpbox VM.