Stubs out the lab "Closing the Support Information Gap"
TL;DR
-----

Creates a skeleton for a Builders Plan support lab to speed iteration

Details
-------

Starts a new Builders Plan support lab implementation so that the
files needed to the Instruqt -> Vendor Portal automation to work can
be on `main` early in the development process. This enables faster
iteration by avoiding the need to pass in the `icp_branch` parameter
and enter the lab at certain URLs.