Adds the "Avoiding Installation Pitfalls" lab (#273)
TL;DR
-----

Implements a preflight lab for the Builders Plan


Details
-------

Provides a lab that teaches how to use preflight checks as part
of an application distributed as a Helm chart using the Replicated
Platform (a.k.a. a Builders Plan app). The learner is brought
through the following steps:

1. Learning what a preflight check is and adding a simple preflight
   check for the Kubernetes verison
2. Expanding the preflight checks to look at cluster resources like
   CPU and Memory.
3. Incorporating the preflight checks into their Helm chart.
4. Releasing a new version of the chart containiner the preflight
   checks
5. Creating a new customer on the Replicated Vendor Portal and 
   running the preflight checks as that customer.
6. Installing the application as that customer.

The lab uses the Bitnami chart for the Harbor image registry and
checks the Kubernetes version, CPU requirements, and memory 
requirements as documented by VMware. The learner first experiences
failed preflights, then the lab setup adds capacity to the cluster so
they can see the preflights pass and complete an install.

To simulate a customer adding capacity to their cluster, the lab
sandbox manipulate the number of nodes in the cluster. At track setup
time, two of the three nodes of the cluster are delated. Between the
failed preflight check and the successful check, lifecycle scripts
for those two nodes restart the K3s service so they will rejoin the
cluster.

The lab does not show preflight checks running as a Helm pre-install
hook at this time. I understand that may not be part of the launch
experience and wanted to make sure I didn't reach something that
will be going away.

Note
----

To try it out, you must run it through [Avoiding Installation Pitfalls](https://play.instruqt.com/embed/replicated/tracks/avoiding-installation-pitfalls?token=em_gJjtIzzTTtdd5RFG&icp_branch=feature-helm-preflights) to 
assure the integration between lab setup and the Vendor Portal
works correctly.