Modernizes KOTS labs with new Instruqt capabilities (#281)
TL;DR
-----

Makes the main KOTS labs more user friendly base on new patterns

Details
-------

I reviewed all the labs on the template invite for KOTS Labs after receiving
some non-specific feedback that they were "broken" and that the issue were in
the "later labs". During my run through, I realized there were some areas for
quick wins in improving the usability of the labs, to wit:

* Labs didn't have meaningful icons.
* All of the labs were written before variables could be embedded in the
  instructions and had convoluted instructions around authentication because of
  it.
* The Replicated CLI should have been updated to use TMUX
* Some `website` tabs for connecting to the Admin Console could be `service`
  tabs instead.
* A few labs called the Admin Console "KotsAdm".
* Some screenshots were confusingly out of date. I may not have caught all of
  these.

I also found a bug in the last challenge of the airgap lab that didn't affect
the outcome but was offputting. It was likely introduced by image repaving and
I updated the Packer template to fix it.

I wasn't able to reproduce any blocking bugs. But the lab experience should be
improved by this work. The KOTS labs still need an overhaul to be more
story-driven exercies vs. tours of features, and could use a more consistent
voice. We're going to start that work soon. These changes will help in the mean
time.