Adds "Closing the Support Information Gap" lab (#277)
TL;DR
-----

Implements a support lab for the builders experience

Details
-------

Provides a lab that teaches how to suport your customer with the
Replicted Platform builders experience. Takes the learned through
the following steps for their builders plan app:

1. Using Instance Insights to gain situatiional awareness for a
   customer issue.
2. Learning what a support bundle is and how to define and
   collect a simple one.
3. Expanding the support build to be a fairly robust bundle for
   application.
4. Incorporating the support bundle specification into their Helm
   chart.
5. Releasing a new version of the chart containiner the support
   bunlde spec.
6. Upgrading the customer's instance and using the in-cluster
   support bundle spec from the Helm chart to collect a support
   bundle.
7. Making a support bundle definition available from a URI and
   using that to assure an up-to-date support bundle spec
   indepdent of new the application release.

The lab uses the Bitnami chart for the Harbor image registry that
is use in other labs. It includes the changes to the Harbor chart
made in those previous labs. It uses `kubectl rollout pause` and
`kubectl delete pod` to cause a simulated outage.