Fixes release promotion instructions
TL;DR
-----

Update release promotion instructions to use match existing sequence
numbers

Details
-------

Up until recently every Builders experience lab started with an empty
release of the application. This was kind of accidental: the lab
automation creates a release for the application, and those labs didn't
really need one. It worked out OK since the Vendor Portal allowed the
empty release and the lab instructions took the release into account.

This all changed when the Vendor portal team tighten the constraints on
releases, at least for Builders experience applications. The empty
release doesn't have a Helm chart in it. So no release is created and
the automation silently failed. The labs continued to work since the
release is created near the end and everything the labs need was in
place.

What did fail was copy/paste of the lab instructions. The instructions
assumed the empty release was there and once it wasn't pain ensued.
Learners had to copy, paste, and update or just got frustrated with
failing commands. To provide them a remedy, this change updates the
instructions to account for the empty release not existing. This is how
the labs were intended to run from the beginning, starting with zero
releases.

Of course this mostly fixes the symptom and is only a short term fix.
The longer term fix is twofold:

1. Correct the automation to not create a release if the vendor track
   setup doesn't provide a directory for it.
2. Allow the track author to specify a version number for the release
   when one is created.

The latter may not be an obvious requirement, but it's needed since the
current release creation hardcodes the version number `0.1.0`. That
won't work with Builders releases since the application version needs to
match the chart version and not all charts being released to support a
lab will be version `0.1.0`.  I discovered this latter part when trying
a quick fix of dropping the Harbor Helm chart into a directory and
specifying that directory to create the release.