Frames up compatibility lab
TL;DR
-----

Creates the framing for a new lab on compatibility testing

Details
-------

Lays the groundwork for a new compability testing lab to avoid having to
set unnatural variables to iterate on the lab. @beck-at-replicated has
built out text for a couple of the challenges and the vendor setup is
stolen from the other labs in the Builders series. Also steals the track
setup from the "Closing the Support Information Gap" since it need to do
all of that setup plus the work of the solve scripts in that lab.