Barebones Admin Console track for automation (#306)
TL;DR
-----

Creates a barebones lab for addding the admin console to Harbor

Details
-------

Provides a bare minimum to distinguish the new lab I'm working on so I
can get the automation to see it and be able to work without setting
customer URL parameters. Lab is in maintenance so should be fine.

Goal of the lab is to shift from Helm-based to KOTS-based distribution
for the Harbor Helm application that's used in the Builders experience
labs. We're looking for examples for this to support docs and demos, so
a lab felt like a good place to start.