Quick Hit Lab for CMX (initial skeleton) (#314)
TL;DR
-----

Sketches out an initial skeleton for a fast CMX lab

Details
-------

Creates the track structure and first challenge for a quick lab on CMX to
use at Kubecon. This PR just gets those files in place to simplify working
with the lab setup as the lab gets full developed.