Deprecates and Removes Old-Style Labs (#319)
TL;DR
-----

Removes redundant labs and tags deprecated ones

Details
-------

Tries to avoid using out-of-date, less maintained, and superceded
labs by:

* Removing original "Application Preflights" lab as it has been superceded by
  "Avoding Installation Pitfalls"
* Removing the three support bundle labs that have been superceded by "Closing
  the Support Information Gap"
* Removing the "Deploying a Helm Based Application with Replicated" which has
  been superceded by "Distributing Your Application with Replicated"
* Marking several labs deprecated that will be superceded by new labs or which
  need to be modernized for the current state of the platform

Also tags a few labs that depend on or show case premium features to highlight
the difference. This aligns with the `builders-plan` tag on labs that cover
features available to those customers.