Polishes up GitHub actions (#321)
TL;DR
-----

Provides human names to GitHub workflows and tidies up a a bit

Details
-------

* Provides real titles for all workflows
* Assure that setup only gets built when it has changes, this 
  lines up with the deployment process as well
* Removes Lab 11 workflow that didn't appear to do anything
  since we don't have a Lab 11 or any other monitoring lab

Caveat
------

There's still a question about whether the way dev tests runs is
the best approach. Currently every dev variant of a current prod
track is tested rather than just the one the author is working on.
This may be the best appraoch since they may be working on multiple
tracks at once, but I'm not 100% sure.