Scaffolds the track Protecting Your Assets with Replicated (#326)
TL;DR
-----

Creates the base structure for a lab on licensing

Details
-------

Provides a barebones lab named "Protecting Your Assets with Replicated" that
take users through the Replicated license and things you can do with it. This
makes it easier to work with the in progress lab including, most critically,
allowing `instruqt track test` to run.