Scaffolds and outlines "Delivering Your Application as a Kubernetes Appliance" (#329)
TL;DR
-----

Frames up the Embedded Cluster lab

Details
-------

Provides an initial outline of and overview lab for the Embedded Cluster. Each
challenge explains what that step of the lab is about and has a full set of
lifecycle scripts with an extremely minimal implementation. Track setup scripts
are not yet implemented.