Corrects merge issue for header download
TL;DR
-----

Fixes an issue with were the branch header was being downloaded

Details
-------

The last updates the template got merged in with the header download in
the track setup script referring to the branch. Once the branch was
deleted it couldn't be downloaded any longer. This change fixes that.