Updates Helm-oriented labs to use Slackernews (#322)
TL;DR
-----

Switches labs that used Harbor to Slackernews

Details
-------

In preparation for writing additional labs that showcase the 
evolution of the Replicated Platform in 2023, I had to make a 
decision: Should I continue with Harbor or shift to Slackernews?

The advantage to using Harbor was that it's a complex application
and showcased that we could take an application built by another 
company (VMware) and packaged by a different team in that
company (Bitnami) and easily distribute it with Replicated.

The disadvantage is that we could get caught off guard by 
changes to the upstream application and chart that would mean 
the labs didn't feel right. We mostly ran into this around
changes to version numbers that made use have to limit the
version we used.

Using Slackernews, on the other hand, carries forward the brand
decision we've made to use it as the best sample application for
Replicated. It also reduces the possiblity of being surprised
by upstream changes. In the end, it seemed like to right call
to switch now before adding additional labs.

This change updates the following three labs:

* Distributing Your Application with Replicated
* Avoiding Installation Pitfalls
* Closing the Support Information Gap

to use the Slackernews application. One lesson from the use of
Harbor that I applied was that these chart versions start with 
their own numbering rather than building on the upstream version
number.

The flow of each lab remains the same, only the application used
has changed.