Adds Embedded Cluster lab (#346)
TL;DR
=====

Provides "Delivering Your Application as a Kubernetes Appliance" lab

Details
=======

Implements a lab named "Delivering Your Application as a Kubernetes
Appliance" that shows how to package and install with Slackernews with the
Replicated Embedded Cluster. 

The lab takes the user through iteratively building the Embedded Cluster
implementation of Slackernews based on the Helm chart. Once they have a
working implementation they install it as a customer then review the new
instance in the Vendor Portal.