Starts an overview lab (#350)
TL;DR
-----

Provides a stub for a new overview-style lab

Details
-------

Implements a stub track for "Meeting the Replicated Platform", a new lab
that vendors can use to discover the Replicated Platform. The lab will
cover some of the same steps as an overview demo to familiarize vendors
with the Platform on their own.