Fixes SDK version fetch for "Distributing" lab (#349)
TL;DR
-----

Uses the SDK version directly in "Distributing" lab

Details
-------

Repairs a regression created by the switch from using the git tag ot the
registry tag to find the SDK version. When the code used the git tag we
had to strip the first character and with the registry version we don't.
This fix makes sure we don't strip it.