Corrects regression in setup scripts that fetch the SDK version (#353)
TL;DR
-----

Stops track setup scripts from stripping the `1` from the SDK version

Details
-------

Fixes a defect in track setup scripts for the "Avoiding Installation
Pitfalls", "Closing the Support Installation Gap", and "Protecting Your
Assets" labs that caused the `1` to be stripped from the SDK version
when the track was created. This defect was introduced when I updated the
function that fetches the SDK version to use the chart registry instead of
GitHub. 

The old code resulted in a version string that included a `v` at the
beginning from the GitHub tag, which was stripped on use. The version tags in
the registry do not have a `v` prefix, and I had lazily removed the `v` prefix
by dropping the first character of the version string. This resulted in the
`1` being stripped once the function changes. This change removes the
expressions that dropped the first character of the version string.

It also assures that track metadata updates Instruqt inserted into several
files in the affected labs is under source control.