Corrects application name for EC lab vendor portal metadata (#354)
TL;DR
-----

Updates metadata for Vendor Portal setup in "Delivering..." lab

Details
-------

Provides a quick fix to a bug that was preventing the "Delivering Your
Application as a Kubernetes Appliance" lab from starting. The track setup was
failing because the vendor metadata used the name "SlackerNews" and the
function to the app slug defaults to "Slackernews".

The natural fix would have been to avoid the default, but there seems to be a
call to `get_app_slug` embedded elsewhere that would blank the value again. I
don't love the inconsistency, but this approach gets the lab running again.

TODO
----

Dig into this problenm more deeply to determine if we should either simplify
the Vendor Portal setup or track down the code that blanks the app slug.