Adapts "Delivering..." to Embedded Cluster chagnes
TL;DR
-----

Updates the "Delivering..." lab for changes to Embedded Cluster

Details
-------

Restores the "Delivering Your Application as a Kubernetes Appliance" lab
to working order. The lab stopped working when changes to the embedded
cluster relocated the `admin.conf` file that a few scripts referenced as
their `KUBECONFIG`.

Also takes advantage of the new(ish) `--admin-console-password` flag to
set the Admin Console password in in the solve script for the fifth
challenge. The earlier script used KOTS to set the password, but
streamlining it to the expected approach feels a lot cleaner and
future-proof.