Adjusts caching on the EC version to avoid errors
TL;DR
-----

Updates the function for fetching the EC version to check the right
variable when catching

Details
-------

We were getting some download errors on the Embedded Cluster. It was
sporadic and there were a couple of options. The first was GitHub rate
limits, which may still be a factor. The second was that the version was
either not being fetched or being fetched incorrectly.

Investigating the second option, I realized that the code that checked
the runtime variable for the version was checking the wrong variable.
This change checks the right variable there. I'm not 100% sure this will
be a lasting fix to the issue, but it was a bug and it was in the right
area.

The code also updates the track checksum (since it changes with every
push) and fixes an error where the config help text had some variable
interpolation problems.