Adapts EC lab for preflight and default updates (#363)
TL;DR
-----

Corrects the Embedded Cluster lab to address changes in Embedded
Cluster preflights and Vendor Portal defaults


Details
--------

Skip host preflights by default in the Embedded Cluster lab. This is
*required* because of an incompatibility between the check for a
wildcard DNS record and the Instruqt infrastructure. Instruqt has a
wildcard DNS entry that causes the preflight to fail, and we need to
have a lab for Embedded Cluster. Since I didn't want to teach the bad
habit of ignoring preflight failures, I had to [add an environment
variable to the Embedded Cluster
CLI](replicatedhq/embedded-cluster#2364) and use it to skip the
preflights.

While I was testing the change, I realized we had issues with
inconsistencies in the defaults in the `replicated` CLI that meant
some of our customer creation steps in lab setup were failing. This
change addresses those issues.

Also removes the KOTS CLI installation process that isn't needed any
longer.